Update README.md to fix the TOC (#945)
* Update README.md to fix the TOC

Fix the table of contents in the README.md, especially fix all the links.

I used doctoc (https://github.com/thlorenz/doctoc) to generate the table of contents and copied it to the Markdown file.

* Shorten the TOC in the README.md

Shorten the table of contents to be on a single page.

* Clean up README by removing redundant links

Removed unnecessary sub-sections from the README.